backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==12.0
pydantic==2.11.7
pydantic-settings==2.6.1
boto3==1.34.0
PyJWT[crypto]==2.8.0
python-multipart==0.0.20
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.28.1
moto==4.2.14
botocore==1.34.0